Added modules in videoHandler to reduce redundancy
