+++ frontend/README.md
@@ -8,20 +8,24 @@ npm install
 npm run dev
 ```
 
-
 ---
 
 ### Code formatting / Linting
 
 - To check all the lint errors in your frontend app:
+
 ```bash
 npm run lint
 ```
+
 - To fix all the lint errors in your frontend app:
+
 ```bash
 npm run lint-fix
 ```
+
 - To format your code changes using prettier:
+
 ```bash
 npm run format
 ```
