+++ frontend/src/config.ts
@@ -1,4 +1,4 @@
 export const BACKEND_URL = 'https://backend.aadesh-kulkarni08.workers.dev';
 export const GITHUB_CONTRIBUTOR_URL = 'https://api.github.com/repos/aadeshkulkarni/medium-app/contributors';
 // export const BACKEND_URL = "http://localhost:8787";
-export const FF_ENABLE_AI = true; // Feature flag (temp solution)
+export const FF_ENABLE_AI = true; // Feature flag (temp solution)
