+++ frontend/src/pages/Publish.tsx
@@ -2,7 +2,7 @@ import axios from 'axios';
 import Appbar from '../components/Appbar';
 import { BACKEND_URL, FF_ENABLE_AI } from '../config';
 import { useEffect, useState, useRef } from 'react';
-import ReactQuill,{Quill} from 'react-quill';
+import ReactQuill, { Quill } from 'react-quill';
 import 'react-quill/dist/quill.bubble.css';
 import { toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
@@ -12,9 +12,7 @@ import { useAI } from '../hooks/blog';
 import GenerateAIBtn from '../components/GenerateAIBtn';
 import PublishTags from '../components/PublishTags';
 import { htmlTagRegex } from '../util/string';
-import { videoHandler } from '../util/videoHandler';
-
-
+import { videoHandler,modules } from '../util/videoHandler';
 
 // Register the custom video handler with Quill toolbar
 Quill.register('modules/customToolbar', function (quill: any) {
@@ -63,20 +61,7 @@ const Publish = () => {
     if (e.key === 'Enter') writingPadRef.current?.focus();
   };
 
-  const modules = {
-    toolbar: {
-      container: [
-        [{ 'header': '1' }, { 'header': '2' }, { 'font': [] }],
-        [{ 'list': 'ordered'}, { 'list': 'bullet' }, { 'list': 'check' }], // Add tasklist option
-        ['bold', 'italic', 'underline', 'strike'],
-        [{ 'align': [] }],
-        [{ 'color': [] }, { 'background': [] }],
-        ['code-block'],
-        ['link', 'image', 'video'],
-        ['clean']
-      ],
-    },
-  };
+  
 
   return (
     <>
@@ -109,7 +94,7 @@ const Publish = () => {
           className="tracking-wide text-[#0B1215] font-light"
           value={content}
           onChange={(value) => setContent(htmlTagRegex.test(value) ? '' : value)}
-          modules={modules} 
+          modules={modules}
         ></ReactQuill>
       </div>
       <ToastWrapper />
